start refactoring to support pbs; mostly moving code around
